Saved vehicles can now be spawned even with the "Vehicle Spawner" menu "disabled" (no permissions) as long as the saved vehicle belongs in a vehicle class that has been allowed.
For example. If you do *not* give players access to the vehiclespawner menu, but you still want them to be able to spawn saved vehicles from the "offroad" class, simply give them the "vMenu.VehicleSpawner.OffRoad" permission and they'll still be able to spawn that vehicle using the saved vehicles menu.