Recode player blips.
Blips will now only be handled if the specific player is nearby or if the pause menu is active. If the player is too far away the blips will be hidden on the minimap (main map will still show them). Fixed sprites not being set correctly when the other player is far way, (this was because the game doesn't know that the other player has a vehicle). This is fixed by using decorators on the player ped.