Door locking mechanism change
Originally, the method SetVehicleDoorsLockedForAllPlayers was used to lock car doors. However, whenever a driver was inside the car and a passenger attempted to enter a locked car, the passenger would smash the window, open the door, attempt to get in, and instantly get kicked outside. Menyoo (the client side mod) was able to lock doors without this issue. Diving into his codebase, he uses the method SetVehicleDoorsLocked, where 0 == unlocked, 2 == locked and 4 == childlock (childlock meaning you can't leave when locked, while 2 allows you to leave a locked vehicle but not reenter). Changing the LockOrUnlockDoors method with this new method, and making sure the "double click E to lock/Unlock" methods are changed, the functionality fully works as expected.