Cruise control functionality
Added the ability to turn on cruise control. Basically the same functionality as cruise control in Menyoo trainer, however, some features were added.
Added permission for cruise control option
Added option for cruise control in vehicle options
When turning on cruise control (checkbox) set the cruise control speed
If the vehicle is a helicopter, gets too a low speed, or is no longer in a vehicle, turn cruise control off automatically
If the vehicle is a plane, if the player is not the driver, if the car is in the air or in the water, don't set the speed of the vehicle
An extra addition to this cruise control feature is that it no longer sets the speed constantly to the car when it slows down due to crashing or other factors, but will now make its way back up to the cruise control speed using the vehicles acceleration. If the vehicle is driving the cruise control speed, it will just set the cruise control speed. Right now when testing this ingame with the vMenu KM/h / MP/h feature, it will flicker between the set value and one below, this is cause the speed gets set in vehicleoptions. This may be fixed by setting the speed every frame instead, but may be more costly. What does happen on "every frame" is the ability to up or lower the cruise control speed using the vehiclebrake and vehicleaccelerate key, upping it by 2KM/h